added stars
