+++ frontend/src/components/GameShow/GameShow.css
@@ -29,13 +29,23 @@
     border-radius: 3px;
 }
 
-.bottom-half {
-    height: 40vh;
+.game-user-info {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center;
+    gap: 10px;
+    min-height: 60vh;
+    flex-grow: 1;
+}
+
+.reviews {
     width: 100%;
     display: flex;
     flex-direction: column;
-    justify-content: center;
     align-items: center;
+    gap: 10px;
+    margin-top: 150px;
 }
 
 .game-info {
